simplify app, start items, split up networking (#10)

+++ src/modules/physics/components.hpp
@@ -42,8 +42,8 @@ struct PhysicsSizeComponent
 
 struct VelocityComponent
 {
-  float x = 0.0f;
-  float y = 0.0f;
+  int x = 0;
+  int y = 0;
 };
 
 // specific for actor-actor collisions
